Divide classes into packages
Added package duke.util and duke.task
