+++ src/main/java/TiTi/task/Task.java
@@ -1,3 +1,6 @@
+package TiTi.task;
+
+
 public class Task {
     protected String description;
     protected boolean isDone;
