+++ src/main/java/TiTi/util/Parser.java
@@ -1,4 +1,12 @@
+package TiTi.util;
+
+import TiTi.task.Task;
+import TiTi.task.Event;
+import TiTi.task.Deadline;
+import TiTi.task.ToDo;
+
 import java.util.Scanner;
+
 
 public class Parser {
     private SavedHistory savedHistory;
