+++ src/main/java/TiTi/util/Response.java
@@ -1,0 +1,17 @@
+package TiTi.util;
+
+import TiTi.task.Task;
+
+
+public class Response {
+    public enum Cue {EXIT, LIST, DONE, DELETE, TASKERROR, TODO, DEADLINE,
+            EVENT, MISSINGDESCRIPTION, UNRECOGNISED};
+    public Cue cue;
+    public Task task;
+
+    public Response(Cue cue, Task task) {
+        this.cue = cue;
+        this.task = task;
+    }
+
+}
